+++ Plugin.cs
@@ -9,7 +9,7 @@ namespace SigilADay_julianperge
 	{
 		public const string PluginGuid = "julianperge.inscryption.sigiladay";
 		private const string PluginName = "SigilADay_julianperge";
-		private const string PluginVersion = "1.3.0";
+		private const string PluginVersion = "1.4.0";
 
 		internal static ManualLogSource Log;