+++ sigils/Valkyrie.cs
@@ -16,11 +16,11 @@ namespace SigilADay_julianperge
 			Texture2D defaultTexture =
 				SigilUtils.LoadTextureFromResource(Resources.card_norse_warrior);
 
-			const string name = "NorseWarrior";
+			const string warriorName = "NorseWarrior";
 			const string displayName = "Norse Warrior";
 			const string desc = "Odin would be proud!";
 
-			NewCard.Add(name, displayName, 1, 1,
+			NewCard.Add(warriorName, displayName, 1, 1,
 				null, CardComplexity.Simple, CardTemple.Nature,
 				desc, bloodCost: 1, defaultTex: defaultTexture, abilities: new List<Ability>() { Ability.Brittle }
 			);
@@ -32,7 +32,8 @@ namespace SigilADay_julianperge
 
 			const string rulebookName = "Flight of the Valkyrie";
 			const string rulebookDescription =
-				"When [creature] is played, fill all open slots on your side of the field with 1/1 Norse Warriors.";
+				"When [creature] is played, fill all open slots on your side of the field with 1/1 Norse Warriors. " +
+				"A Norse Warrior is defined as: 1 Power, 1 Health, Brittle";
 
 			// setup ability
 			AbilityInfo info = SigilUtils.CreateInfoWithDefaultSettings(rulebookName, rulebookDescription, true, 3);